Refactor
Apply latest changes from Spring Boot 3.1.0-SNAPSHOT and centralize
container definitions to be used for test and dev mode.
