--- write-app/src/test/java/com/salaboy/dapr/javaapp/DaprSidecarContainer.java
@@ -1,16 +1,0 @@
-package com.salaboy.dapr.javaapp;
-
-import org.testcontainers.containers.GenericContainer;
-
-public class DaprSidecarContainer extends GenericContainer<DaprSidecarContainer> {
-
-    DaprSidecarContainer() {
-        super("daprio/daprd:edge");
-        withCommand("./daprd",
-                "-app-id", "write-app",
-                "-placement-host-address", "placement:50006",
-                "--dapr-listen-addresses=0.0.0.0",
-                "-components-path", "/components");
-        withExposedPorts(50001);
-    }
-}
