--- write-app/src/test/java/com/salaboy/dapr/javaapp/RedisContainer.java
@@ -1,11 +1,0 @@
-package com.salaboy.dapr.javaapp;
-
-import org.testcontainers.containers.GenericContainer;
-
-public class RedisContainer extends GenericContainer<RedisContainer> {
-
-    RedisContainer() {
-        super("redis:alpine");
-        withExposedPorts(6379);
-    }
-}
